test: split live capacity diagnostics
